+++ mex/extractors/publisher/main.py
@@ -2,7 +2,6 @@ from mex.common.cli import entrypoint
 from mex.common.models import AnyMergedModel, ItemsContainer
 from mex.extractors.pipeline import asset, run_job_in_process
 from mex.extractors.publisher.extract import get_merged_items
-from mex.extractors.publisher.filter import filter_merged_items
 from mex.extractors.settings import Settings
 from mex.extractors.sinks import load
 
@@ -10,9 +9,7 @@ from mex.extractors.sinks import load
 @asset(group_name="publisher")
 def extract_and_filter_merged_items() -> ItemsContainer[AnyMergedModel]:
     """Get merged items from mex-backend and filter them by allow-list."""
-    items = get_merged_items()
-    filtered = list(filter_merged_items(items))
-    return ItemsContainer[AnyMergedModel](items=filtered)
+    return ItemsContainer[AnyMergedModel](items=get_merged_items())
 
 
 @asset(group_name="publisher")